+++ src/openforms/submissions/logic/actions.py
@@ -76,6 +76,8 @@ class ActionOperation:
         configuration: FormioConfigurationWrapper,
         submission: Submission,
         initial_data: FormioData,
+        *,
+        original_input_data: FormioData,
     ) -> DataMapping | None:
         """
         Return a mapping [name/path -> new_value] with changes that are to be
@@ -101,6 +103,9 @@ class PropertyAction(ActionOperation):
     def apply(
         self, step: SubmissionStep, configuration: FormioConfigurationWrapper
     ) -> None:
+        # handled directly inside eval
+        if self.property == "hidden":
+            return None
         if self.component not in configuration:
             return None
         component = configuration[self.component]
@@ -112,20 +117,32 @@ class PropertyAction(ActionOperation):
         configuration: FormioConfigurationWrapper,
         submission: Submission,
         initial_data: FormioData,
+        *,
+        original_input_data: FormioData,
     ) -> DataMapping | None:
         # To avoid doing unnecessary work, only apply clear-on-hide logic for components
-        # for which their action sets the "hidden" property to True.
-        if not (self.property == "hidden" and self.value):
+        # for which their action touches the "hidden" property.
+        if self.property != "hidden":
             return None
 
-        # Note that this can happen when the action applies to a component of a
-        # different step than we are currently evaluating. We don't have to do
+        # we now know that we're making the component visible or hidden, but we're not
+        # guaranteed that we're actually changing the visibility state at all! For that
+        # we need to check at the current component configuration, and that requires
+        # first checking if the component is in the current step's configuration. The
+        # component *can* be absent if it refers to a component in another step than
+        # the one we are currently evaluating. We don't have to do
         # anything in this case, because the clear-on-hide behaviour of that
         # component will be handled once the user has reached that step.
         if self.component not in configuration:
             return None
 
         component = configuration[self.component]
+        should_be_hidden = self.value is True
+
+        # apply the state mutation immediately, so that it's visible to follow up
+        # actions and rules operating on the same component. This obsoletes the
+        # apply method/side-effects.
+        component[self.property] = self.value
 
         # Process the visibility of the component. We want to process the component
         # itself, not try to iterate over its children, so we create a 'fake'
@@ -135,7 +152,8 @@ class PropertyAction(ActionOperation):
             context,
             configuration,
             initial_data=initial_data,
-            parent_hidden=True,
+            parent_hidden=should_be_hidden,
+            original_input_data=original_input_data,
         )
         return None
 
@@ -217,6 +235,8 @@ class VariableAction(ActionOperation):
         configuration: FormioConfigurationWrapper,
         submission: Submission,
         initial_data: FormioData,
+        *,
+        original_input_data: FormioData,
     ) -> DataMapping:
         with log_errors(self.value, self.rule):
             return {self.variable: jsonLogic(self.value, context.data)}
@@ -326,6 +346,8 @@ class SynchronizeVariablesAction(ActionOperation):
         configuration: FormioConfigurationWrapper,
         submission: Submission,
         initial_data: FormioData,
+        *,
+        original_input_data: FormioData,
     ) -> DataMapping | None:
         configuration = submission.total_configuration_wrapper
         if self.source_variable not in configuration:
@@ -349,6 +371,8 @@ class ServiceFetchAction(ActionOperation):
         configuration: FormioConfigurationWrapper,
         submission: Submission,
         initial_data: FormioData,
+        *,
+        original_input_data: FormioData,
     ) -> DataMapping:
         var = self.rule.form.formvariable_set.get(key=self.variable)
         with log_errors({}, self.rule):  # TODO proper error handling
@@ -392,6 +416,8 @@ class EvaluateDMNAction(ActionOperation):
         configuration: FormioConfigurationWrapper,
         submission: Submission,
         initial_data: FormioData,
+        *,
+        original_input_data: FormioData,
     ) -> DataMapping | None:
         # Mapping from form variables to DMN inputs
         dmn_inputs = {