+++ src/openforms/submissions/tests/form_logic/test_submission_logic.py
@@ -1132,6 +1132,430 @@ class CheckLogicSubmissionTest(SubmissionsMixin, APITestCase):
         self.assertNotIn("date", data["step"]["data"])
 
 
+@tag("gh-6005")
+class MultipleRulesTargettingSameComponentVisibilityTests(
+    SubmissionsMixin, APITestCase
+):
+    def test_first_rule_triggers_second_does_not_must_not_clear_value(self):
+        form = FormFactory.create(
+            generate_minimal_setup=True,
+            formstep__form_definition__configuration={
+                "components": [
+                    {
+                        "type": "radio",
+                        "key": "radio",
+                        "values": [
+                            {"value": "a", "label": "A"},
+                            {"value": "b", "label": "B"},
+                        ],
+                    },
+                    {
+                        "type": "textfield",
+                        "key": "show-when-a",
+                        "hidden": True,
+                    },
+                ]
+            },
+        )
+        form_step = form.formstep_set.get()
+        # just a single rule does not reproduce the issue (!).
+        # Rule 1: show the textfield when 'a' is selected in the radio
+        show_textfield_action = {
+            "formStep": None,
+            "component": "show-when-a",
+            "action": {
+                "type": "property",
+                "property": {"value": "hidden", "type": "bool"},
+                "value": {},
+                "state": False,
+            },
+        }
+        FormLogicFactory.create(
+            form=form,
+            json_logic_trigger={"==": [{"var": "radio"}, "a"]},
+            actions=[show_textfield_action],
+        )
+        # Rule 2: show the textfield when a non-sense value is selected in the radio
+        # (never triggers!)
+        FormLogicFactory.create(
+            form=form,
+            json_logic_trigger={"==": [{"var": "radio"}, "nonsense-value"]},
+            actions=[show_textfield_action],
+        )
+        submission = SubmissionFactory.create(form=form)
+        self._add_submission_to_session(submission)
+        logic_check_endpoint = reverse(
+            "api:submission-steps-logic-check",
+            kwargs={
+                "submission_uuid": submission.uuid,
+                "step_uuid": form_step.uuid,
+            },
+        )
+
+        # this assumes there has been a call before with `radio: a` that results in
+        # show-when-a becoming visible, after which the user enters a value for that
+        # field.
+        response = self.client.post(
+            logic_check_endpoint,
+            {"data": {"radio": "a", "show-when-a": "do-not-clear-me"}},
+        )
+
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+
+        self.assertFalse(
+            data["step"]["formStep"]["configuration"]["components"][1]["hidden"]
+        )
+        self.assertEqual(data["step"]["data"], {})
+
+    def test_second_rule_triggers_first_does_not_must_not_clear_value(self):
+        form = FormFactory.create(
+            generate_minimal_setup=True,
+            formstep__form_definition__configuration={
+                "components": [
+                    {
+                        "type": "radio",
+                        "key": "radio",
+                        "values": [
+                            {"value": "a", "label": "A"},
+                            {"value": "b", "label": "B"},
+                        ],
+                    },
+                    {
+                        "type": "textfield",
+                        "key": "show-when-a",
+                        "hidden": True,
+                    },
+                ]
+            },
+        )
+        form_step = form.formstep_set.get()
+        # just a single rule does not reproduce the issue (!).
+        # Rule 1: show the textfield when a non-sense value is selected in the radio
+        # (never triggers!)
+        show_textfield_action = {
+            "formStep": None,
+            "component": "show-when-a",
+            "action": {
+                "type": "property",
+                "property": {"value": "hidden", "type": "bool"},
+                "value": {},
+                "state": False,
+            },
+        }
+        FormLogicFactory.create(
+            form=form,
+            json_logic_trigger={"==": [{"var": "radio"}, "nonsense-value"]},
+            actions=[show_textfield_action],
+        )
+        # Rule 2: show the textfield when 'a' is selected in the radio
+        FormLogicFactory.create(
+            form=form,
+            json_logic_trigger={"==": [{"var": "radio"}, "a"]},
+            actions=[show_textfield_action],
+        )
+        submission = SubmissionFactory.create(form=form)
+        self._add_submission_to_session(submission)
+        logic_check_endpoint = reverse(
+            "api:submission-steps-logic-check",
+            kwargs={
+                "submission_uuid": submission.uuid,
+                "step_uuid": form_step.uuid,
+            },
+        )
+
+        response = self.client.post(
+            logic_check_endpoint,
+            {"data": {"radio": "a", "show-when-a": "do-not-clear-me"}},
+        )
+
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+
+        self.assertFalse(
+            data["step"]["formStep"]["configuration"]["components"][1]["hidden"]
+        )
+        self.assertEqual(data["step"]["data"], {})
+
+    def test_rules_flip_from_hidden_to_visible_state_should_not_clear_value(self):
+        form = FormFactory.create(
+            generate_minimal_setup=True,
+            formstep__form_definition__configuration={
+                "components": [
+                    {
+                        "type": "radio",
+                        "key": "radio",
+                        "values": [
+                            {"value": "a", "label": "A"},
+                            {"value": "b", "label": "B"},
+                        ],
+                    },
+                    {
+                        "type": "checkbox",
+                        "key": "checkbox",
+                    },
+                    {
+                        "type": "textfield",
+                        "key": "hide-when-a-but-show-when-checkbox-checked",
+                        "hidden": False,
+                    },
+                    {
+                        "type": "fieldset",
+                        "key": "fieldset",
+                        "hidden": False,
+                        "components": [
+                            {
+                                "type": "textfield",
+                                "key": "nestedTextfield",
+                                "hidden": False,
+                            }
+                        ],
+                    },
+                ]
+            },
+        )
+        form_step = form.formstep_set.get()
+
+        def _build_visibility_action(key: str, make_hidden: bool):
+            return {
+                "formStep": None,
+                "component": key,
+                "action": {
+                    "type": "property",
+                    "property": {"value": "hidden", "type": "bool"},
+                    "value": {},
+                    "state": make_hidden,
+                },
+            }
+
+        # Hide (and clear) the textfield when 'a' is selected in the radio
+        FormLogicFactory.create(
+            form=form,
+            json_logic_trigger={"==": [{"var": "radio"}, "a"]},
+            actions=[
+                _build_visibility_action(
+                    key="hide-when-a-but-show-when-checkbox-checked", make_hidden=True
+                ),
+                _build_visibility_action(key="fieldset", make_hidden=True),
+            ],
+        )
+        # Expected to trigger: the value of the textfield gets cleared because of the
+        # above rule.
+        # NOTE: this rule is illegal in 3.5.x due to the cyclic reference. When
+        # backporting, make sure it targets another component.
+        FormLogicFactory.create(
+            form=form,
+            json_logic_trigger={
+                "==": [{"var": "hide-when-a-but-show-when-checkbox-checked"}, ""]
+            },
+            actions=[
+                {
+                    "formStep": None,
+                    "component": "hide-when-a-but-show-when-checkbox-checked",
+                    "action": {
+                        "type": "property",
+                        "property": {"value": "validate.required", "type": "bool"},
+                        "value": {},
+                        "state": True,
+                    },
+                }
+            ],
+        )
+        # Show the textfield when the checkbox is checked
+        FormLogicFactory.create(
+            form=form,
+            json_logic_trigger={"var": "checkbox"},
+            actions=[
+                _build_visibility_action(
+                    key="hide-when-a-but-show-when-checkbox-checked", make_hidden=False
+                ),
+                _build_visibility_action(key="fieldset", make_hidden=False),
+            ],
+        )
+
+        submission = SubmissionFactory.create(form=form)
+        self._add_submission_to_session(submission)
+        logic_check_endpoint = reverse(
+            "api:submission-steps-logic-check",
+            kwargs={
+                "submission_uuid": submission.uuid,
+                "step_uuid": form_step.uuid,
+            },
+        )
+
+        response = self.client.post(
+            logic_check_endpoint,
+            {
+                "data": {
+                    "radio": "a",
+                    "checkbox": True,
+                    "hide-when-a-but-show-when-checkbox-checked": "do-not-clear-me",
+                    "nestedTextfield": "do-not-clear-me",
+                }
+            },
+        )
+
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+
+        with self.subTest("textfield"):
+            textfield_component = data["step"]["formStep"]["configuration"][
+                "components"
+            ][2]
+            self.assertFalse(textfield_component["hidden"])
+            self.assertTrue(textfield_component["validate"]["required"])
+
+        with self.subTest("fieldset"):
+            fieldset_component = data["step"]["formStep"]["configuration"][
+                "components"
+            ][3]
+            self.assertFalse(fieldset_component["hidden"])
+            self.assertFalse(fieldset_component["components"][0]["hidden"])
+
+        with self.subTest("data mutations"):
+            self.assertEqual(data["step"]["data"], {})
+
+    def test_rules_flip_from_hidden_to_visible_state_should_not_clear_value_inverse(
+        self,
+    ):
+        form = FormFactory.create(
+            generate_minimal_setup=True,
+            formstep__form_definition__configuration={
+                "components": [
+                    {
+                        "type": "radio",
+                        "key": "radio",
+                        "values": [
+                            {"value": "a", "label": "A"},
+                            {"value": "b", "label": "B"},
+                        ],
+                    },
+                    {
+                        "type": "checkbox",
+                        "key": "checkbox",
+                    },
+                    {
+                        "type": "textfield",
+                        "key": "hide-when-a-but-show-when-checkbox-checked",
+                        "hidden": True,
+                    },
+                    {
+                        "type": "fieldset",
+                        "key": "fieldset",
+                        "hidden": True,
+                        "components": [
+                            {
+                                "type": "textfield",
+                                "key": "nestedTextfield",
+                                "hidden": False,
+                            }
+                        ],
+                    },
+                ]
+            },
+        )
+        form_step = form.formstep_set.get()
+
+        def _build_visibility_action(key: str, make_hidden: bool):
+            return {
+                "formStep": None,
+                "component": key,
+                "action": {
+                    "type": "property",
+                    "property": {"value": "hidden", "type": "bool"},
+                    "value": {},
+                    "state": make_hidden,
+                },
+            }
+
+        # Hide (and clear) the textfield when 'a' is selected in the radio
+        FormLogicFactory.create(
+            form=form,
+            json_logic_trigger={"==": [{"var": "radio"}, "a"]},
+            actions=[
+                _build_visibility_action(
+                    key="hide-when-a-but-show-when-checkbox-checked", make_hidden=True
+                ),
+                _build_visibility_action(key="fieldset", make_hidden=True),
+            ],
+        )
+        # Expected to trigger: the value of the textfield gets cleared because of the
+        # above rule.
+        # NOTE: this rule is illegal in 3.5.x due to the cyclic reference. When
+        # backporting, make sure it targets another component.
+        FormLogicFactory.create(
+            form=form,
+            json_logic_trigger={
+                "==": [{"var": "hide-when-a-but-show-when-checkbox-checked"}, ""]
+            },
+            actions=[
+                {
+                    "formStep": None,
+                    "component": "hide-when-a-but-show-when-checkbox-checked",
+                    "action": {
+                        "type": "property",
+                        "property": {"value": "validate.required", "type": "bool"},
+                        "value": {},
+                        "state": True,
+                    },
+                }
+            ],
+        )
+        # Show the textfield when the checkbox is checked
+        FormLogicFactory.create(
+            form=form,
+            json_logic_trigger={"var": "checkbox"},
+            actions=[
+                _build_visibility_action(
+                    key="hide-when-a-but-show-when-checkbox-checked", make_hidden=False
+                ),
+                _build_visibility_action(key="fieldset", make_hidden=False),
+            ],
+        )
+
+        submission = SubmissionFactory.create(form=form)
+        self._add_submission_to_session(submission)
+        logic_check_endpoint = reverse(
+            "api:submission-steps-logic-check",
+            kwargs={
+                "submission_uuid": submission.uuid,
+                "step_uuid": form_step.uuid,
+            },
+        )
+
+        response = self.client.post(
+            logic_check_endpoint,
+            {
+                "data": {
+                    "radio": "a",
+                    "checkbox": True,
+                    "hide-when-a-but-show-when-checkbox-checked": "do-not-clear-me",
+                    "nestedTextfield": "do-not-clear-me",
+                }
+            },
+        )
+
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+
+        with self.subTest("textfield"):
+            textfield_component = data["step"]["formStep"]["configuration"][
+                "components"
+            ][2]
+            self.assertFalse(textfield_component["hidden"])
+            self.assertTrue(textfield_component["validate"]["required"])
+
+        with self.subTest("fieldset"):
+            fieldset_component = data["step"]["formStep"]["configuration"][
+                "components"
+            ][3]
+            self.assertFalse(fieldset_component["hidden"])
+            self.assertFalse(fieldset_component["components"][0]["hidden"])
+
+        with self.subTest("data mutations"):
+            self.assertEqual(data["step"]["data"], {})
+
+
 def is_valid_expression(expr: dict):
     try:
         JsonLogicValidator()(expr)